refactor: change views names to better describe what are they
