+++ data/gradience.gresource.xml
@@ -5,9 +5,9 @@
     <file>presets/adwaita-dark.json</file>
     <file>presets/pretty-purple.json</file>
 
-    <file preprocess="xml-stripblanks">ui/error.ui</file>
+    <file preprocess="xml-stripblanks">ui/error_list_row.ui</file>
     <file preprocess="xml-stripblanks">ui/palette_shades.ui</file>
-    <file preprocess="xml-stripblanks">ui/option.ui</file>
+    <file preprocess="xml-stripblanks">ui/option_row.ui</file>
     <file preprocess="xml-stripblanks">ui/window.ui</file>
     <file preprocess="xml-stripblanks">ui/app_type_dialog.ui</file>
     <file preprocess="xml-stripblanks">ui/custom_css_group.ui</file>
@@ -16,10 +16,10 @@
     <file preprocess="xml-stripblanks">ui/preset_row.ui</file>
     <file preprocess="xml-stripblanks">ui/presets_manager_window.ui</file>
     <file preprocess="xml-stripblanks">ui/plugin_row.ui</file>
-    <file preprocess="xml-stripblanks">ui/welcome.ui</file>
-    <file preprocess="xml-stripblanks">ui/preferences.ui</file>
+    <file preprocess="xml-stripblanks">ui/welcome_window.ui</file>
+    <file preprocess="xml-stripblanks">ui/preferences_window.ui</file>
     <file preprocess="xml-stripblanks">ui/repo_row.ui</file>
-    <file preprocess="xml-stripblanks">ui/no_plugin_pref.ui</file>
+    <file preprocess="xml-stripblanks">ui/no_plugin_window.ui</file>
     <file preprocess="xml-stripblanks">ui/share_window.ui</file>
     <file>style.css</file>
     <file>images/welcome.svg</file>
