+++ data/ui/error_list_row.blp
@@ -1,7 +1,7 @@
 using Gtk 4.0;
 using Adw 1;
 
-template GradienceError : ListBoxRow {
+template GradienceErrorListRow : ListBoxRow {
   Box {
     orientation: vertical;
     margin-top: 6;
