+++ data/ui/meson.build
@@ -1,20 +1,20 @@
 blueprints = custom_target('blueprints',
   input: files(
-    'error.blp',
+    'error_list_row.blp',
     'palette_shades.blp',
-    'option.blp',
+    'option_row.blp',
     'window.blp',
     'app_type_dialog.blp',
     'custom_css_group.blp',
     'presets_manager_window.blp',
-    'preferences.blp',
+    'preferences_window.blp',
     'plugin_row.blp',
-    'welcome.blp',
+    'welcome_window.blp',
     'preset_row.blp',
     'builtin_preset_row.blp',
     'explore_preset_row.blp',
     'repo_row.blp',
-    'no_plugin_pref.blp',
+    'no_plugin_window.blp',
     'share_window.blp',
   ),
   output: '.',
