+++ data/ui/no_plugin_window.blp
