+++ data/ui/option_row.blp
@@ -1,7 +1,7 @@
 using Gtk 4.0;
 using Adw 1;
 
-template GradienceOption : Adw.ActionRow {
+template GradienceOptionRow : Adw.ActionRow {
   [suffix]
   MenuButton warning-button {
     valign: center;
