+++ gradience/main.py
@@ -27,11 +27,11 @@ from material_color_utilities_python import *
 from gi.repository import Gtk, Gdk, Gio, Adw, GLib, Xdp, XdpGtk4
 
 from gradience.ui.window import GradienceMainWindow
-from gradience.ui.welcome import GradienceWelcomeWindow
-from gradience.ui.preferences import GradiencePreferencesWindow
+from gradience.ui.welcome_window import GradienceWelcomeWindow
 from gradience.ui.app_type_dialog import GradienceAppTypeDialog
 from gradience.ui.custom_css_group import GradienceCustomCSSGroup
 from gradience.ui.presets_manager_window import GradiencePresetWindow
+from gradience.ui.preferences_window import GradiencePreferencesWindow
 from gradience.utils.css import load_preset_from_css
 from gradience.utils.utils import to_slug_case, buglog, run_command
 from gradience.utils.preset import Preset, presets_dir
