+++ gradience/ui/option_row.py
@@ -1,4 +1,4 @@
-# option.py
+# option_row.py
 #
 # Change the look of Adwaita, with ease
 # Copyright (C) 2022  Gradience Team
@@ -21,9 +21,9 @@ from gi.repository import Gtk, Gdk, Adw
 from gradience.constants import rootdir
 
 
-@Gtk.Template(resource_path=f"{rootdir}/ui/option.ui")
-class GradienceOption(Adw.ActionRow):
-    __gtype_name__ = "GradienceOption"
+@Gtk.Template(resource_path=f"{rootdir}/ui/option_row.ui")
+class GradienceOptionRow(Adw.ActionRow):
+    __gtype_name__ = "GradienceOptionRow"
 
     color_value = Gtk.Template.Child("color-value")
     text_value = Gtk.Template.Child("text-value")
